refactoring : moved app layout into app layer.
