+++ src/app/layout/ui/AppLayout.tsx
@@ -1,7 +1,7 @@
 import LogoFull from '@assets/logo-full.svg';
 import { useMediaQueryState } from '@shared/lib';
 import { ModeSwitcher } from '@features/mode-switcher';
-import { BottomNavbar, SideNavbar, TopNavbar } from '../../navigation';
+import { BottomNavbar, SideNavbar, TopNavbar } from '@widgets/navigation';
 import Layout from './Layout';
 
 const AppLayout = () => {
